refactor : Validation 확인을 객체 자신이 담당하도록 변경
1. StringCalculator 자신이 수식의 항 개수 유효성 검사 추가
2. StringCalculator 애서 is~ 형태로 유효성 결과를 리턴을 받아 throw 발생 하던 방식을 void check~ 형태로
   메소드에서 조건을 만족하면 throw를 발생하는 방식으로 변경
3. Operand 에서 isOperand를 valueOf 하는 과정에서 유효성 확인이 되기 때문에 삭제
4. Operator 에서 isOperatorSymbol를 valueOf 하는 과정에서 유효성 확인이 되기 때문에 삭제
