+++ src/main/java/calculator/Operand.java
@@ -8,8 +8,13 @@ public class Operand {
     }
 
     public static Operand valueOf(String number) {
-        Float numberFloat = Float.valueOf(number);
-        return new Operand(numberFloat);
+        try {
+            Float numberFloat = Float.valueOf(number);
+            return new Operand(numberFloat);
+        }
+        catch (NumberFormatException e) {
+            throw new IllegalArgumentException(number + " cannot be converted to a number. Probably not a numeric string.");
+        }
     }
 
     public Float getNumber() {
@@ -21,13 +26,4 @@ public class Operand {
         return new Operand(calculatedValue);
     }
 
-    public static boolean isOperand(String number) {
-        try {
-            Float.valueOf(number);
-            return true;
-        }
-        catch (NumberFormatException e){
-            return false;
-        }
-    }
 }
