+++ src/main/java/calculator/StringCalculator.java
@@ -2,15 +2,14 @@ package calculator;
 
 public class StringCalculator {
 
+    private final int MINIMUM_COUNT_FORMULA_ELEMENT = 3;
+
     private String[] formulaElements;
 
     public StringCalculator(String formula) {
-        if(isNullOrEmpty(formula)) {
-            throw new IllegalArgumentException("string is null.");
-        }
-
-        String[] formulaBeforeChecking = formula.split(" ");
-        formulaElements = new FormulaValidation(formulaBeforeChecking).check();
+        checkNullOrEmpty(formula);
+        formulaElements = formula.split(" ");
+        checkElementCountOfInvalid(formulaElements);
     }
 
     public float calculate() throws ArithmeticException {
@@ -22,10 +21,20 @@ public class StringCalculator {
         return result.getNumber();
     }
 
-    private boolean isNullOrEmpty(String str) {
-        return (str == null || str.isEmpty());
+    private void checkNullOrEmpty(String str) {
+        if(str == null || str.isEmpty()) {
+            throw new IllegalArgumentException("string is null.");
+        }
     }
 
+    private void checkElementCountOfInvalid(String[] formulaElements) {
+        int length = formulaElements.length;
+        if (length < MINIMUM_COUNT_FORMULA_ELEMENT || isEven(length)) {
+            throw new IllegalArgumentException("There are errors in the number of operators and operands.");
+        }
+    }
 
-
+    private boolean isEven(int number) {
+        return (number % 2 == 0);
+    }
 }
